requirements.txt
absl-py==2.3.1
accelerate==1.12.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
asttokens==3.0.1
attrs==25.4.0
bitsandbytes==0.48.2
blinker==1.9.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
et_xmlfile==2.0.0
executing==2.2.1
faiss-cpu==1.13.0
filelock==3.19.1
fonttools==4.60.1
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
ipykernel==7.1.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.9.1
kiwisolver==1.4.9
MarkupSafe==2.1.5
matplotlib==3.10.7
matplotlib-inline==0.2.1
mpmath==1.3.0
narwhals==2.12.0
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.2
numpy==2.3.3
ollama==0.6.1
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
parso==0.8.5
pillow==11.3.0
platformdirs==4.5.0
prompt_toolkit==3.0.52
protobuf==6.33.1
psutil==7.1.3
pure_eval==0.2.3
pyarrow==21.0.0
pydantic==2.12.4
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rouge_score==0.1.2
rpds-py==0.29.0
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.16.3
sentence-transformers==5.1.2
setuptools==70.2.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
stack-data==0.6.3
streamlit==1.51.0
sympy==1.13.1
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.22.1
toml==0.10.2
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
wcwidth==0.2.14
wheel==0.45.1
